requirements.txt
django==4.2.11
djangorestframework==3.15.1
boto3==1.34.0  # Para simulação de AWS